merge

+++ SWE2_Micro_ERP_Schnittstelle/src/at/technikum/wien/winterhalder/kreuzriegler/swe2/request/GetInvoicesByContactIdRequest.java
@@ -2,26 +2,27 @@ package at.technikum.wien.winterhalder.kreuzriegler.swe2.request;
 
 public class GetInvoicesByContactIdRequest {
 
-	private int contactId;
+	private long contactId;
 
 	/**
 	 * @param contactId
 	 */
-	public GetInvoicesByContactIdRequest(int contactId) {
+	public GetInvoicesByContactIdRequest(long contactId) {
 		this.contactId = contactId;
 	}
 
 	/**
 	 * @return the contactId
 	 */
-	public int getContactId() {
+	public long getContactId() {
 		return contactId;
 	}
 
 	/**
-	 * @param contactId the contactId to set
+	 * @param contactId
+	 *            the contactId to set
 	 */
-	public void setContactId(int contactId) {
+	public void setContactId(long contactId) {
 		this.contactId = contactId;
 	}
